Remove prompt passing in inherited menus
In previous versions it seems that the prompt was being passed to the
parent menu from menus such as the track menu. This has been causing an
error since the parent menu no longer has a prompt parameter (and the
prompt was going unused besides)

This commit thusly removes the prompt argument from calls of super on
some of the child menus.